feat(ui): Data Marketplace overview header seam + reusable create-drawer hooks
Follow-up to #29839 for the Data Marketplace overview.

- DataMarketplacePage: add a `renderPageHeader` seam so AI mode can replace the
  greeting-banner + search hero with a HeaderShell page header, and render the
  content full-width in that mode. Classic mode is unchanged.
- MarketplaceSearchBar: add a `compact` prop for header-embedded sizing (36px
  control height, no bottom margin) so callers don't override its internals.
- Extract `useDomainCreateDrawer` / `useDataProductCreateDrawer` so the create
  drawers can be opened from any surface; DomainListPage and DataProductListPage
  now consume them instead of inlining the form/submit/drawer setup.
