+++ openmetadata-ui/src/main/resources/ui/src/pages/DataMarketplacePage/DataMarketplacePage.component.tsx
@@ -54,7 +54,18 @@ const normalizeLayout = (l: WidgetConfig[]) =>
     }))
     .sort((a, b) => a.y - b.y);
 
-const DataMarketplacePage = () => {
+interface DataMarketplacePageProps {
+  /**
+   * Optional page-header renderer. When provided (AI mode), it replaces the
+   * default greeting-banner + search hero — the caller's header owns the
+   * title, breadcrumb, search and actions. Omit for the classic hero.
+   */
+  renderPageHeader?: () => ReactNode;
+}
+
+const DataMarketplacePage = ({
+  renderPageHeader,
+}: DataMarketplacePageProps) => {
   const { selectedPersona } = useApplicationStore();
 
   const defaultLayout = useMemo(
@@ -124,22 +135,32 @@ const DataMarketplacePage = () => {
     return <Loader />;
   }
 
+  const gridWrapperClassName = `marketplace-grid-wrapper${
+    renderPageHeader ? ' tw:!max-w-none' : ''
+  }`;
+
   return (
     <div className="tw:h-full tw:overflow-y-auto">
       <div className="tw:mb-8">
-        <div
-          className="marketplace-header-bg"
-          style={
-            { '--marketplace-bg': `url(${marketplaceBg})` } as CSSProperties
-          }>
-          <div className="marketplace-grid-wrapper" dir="ltr">
-            <div className="p-x-box">
-              <MarketplaceGreetingBanner />
-              <MarketplaceSearchBar />
+        {renderPageHeader ? (
+          <div className={gridWrapperClassName} dir="ltr">
+            <div className="p-x-box tw:pt-4">{renderPageHeader()}</div>
+          </div>
+        ) : (
+          <div
+            className="marketplace-header-bg"
+            style={
+              { '--marketplace-bg': `url(${marketplaceBg})` } as CSSProperties
+            }>
+            <div className="marketplace-grid-wrapper" dir="ltr">
+              <div className="p-x-box">
+                <MarketplaceGreetingBanner />
+                <MarketplaceSearchBar />
+              </div>
             </div>
           </div>
-        </div>
-        <div className="marketplace-grid-wrapper" dir="ltr">
+        )}
+        <div className={gridWrapperClassName} dir="ltr">
           <div className="p-x-box">
             <AnnouncementsWidgetV2 widgetKey="announcements" />
           </div>
